requirements.txt
alembic==1.0.7
astroid==2.1.0
autopep8==1.4.3
blinker==1.4
Click==7.0
Flask==1.0.2
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
mysqlclient==1.4.1
Pillow==5.4.1
pkg-resources==0.0.0
pycodestyle==2.4.0
pylint==2.2.2
python-dateutil==2.7.5
python-editor==1.0.3
python-slugify==2.0.1
six==1.12.0
SQLAlchemy==1.2.17
typed-ast==1.2.0
Unidecode==1.0.23
Werkzeug==0.14.1
wrapt==1.11.1
WTForms==2.2.1
